+++ plugins/tasks/ansible/src/main/java/com/walmartlabs/concord/plugins/ansible/ArgUtils.java
@@ -26,7 +26,6 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.Arrays;
 import java.util.Collection;
-import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;