+++ plugins/tasks/resource/src/main/java/com/walmartlabs/concord/plugins/resource/ResourceTask.java
@@ -34,8 +34,6 @@ import java.io.OutputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.concurrent.Callable;
-import java.util.function.Function;
 
 @Named("resource")
 @SuppressWarnings("unused")