+++ server/impl/src/main/java/com/walmartlabs/concord/server/org/inventory/InventoryQueryResource.java
@@ -22,7 +22,7 @@ package com.walmartlabs.concord.server.org.inventory;
 
 import com.walmartlabs.concord.common.validation.ConcordKey;
 import com.walmartlabs.concord.server.OperationResult;
-import com.walmartlabs.concord.server.metrics.WithTimer;
+import com.walmartlabs.concord.server.sdk.metrics.WithTimer;
 import com.walmartlabs.concord.server.org.OrganizationEntry;
 import com.walmartlabs.concord.server.org.OrganizationManager;
 import com.walmartlabs.concord.server.org.ResourceAccessLevel;