+++ server/impl/src/main/java/com/walmartlabs/concord/server/process/pipelines/processors/policy/ProcessTimeoutPolicyApplier.java
@@ -25,7 +25,7 @@ import com.walmartlabs.concord.policyengine.CheckResult;
 import com.walmartlabs.concord.policyengine.PolicyEngine;
 import com.walmartlabs.concord.policyengine.ProcessTimeoutRule;
 import com.walmartlabs.concord.sdk.Constants;
-import com.walmartlabs.concord.server.metrics.InjectCounter;
+import com.walmartlabs.concord.server.sdk.metrics.InjectCounter;
 import com.walmartlabs.concord.server.process.Payload;
 import com.walmartlabs.concord.server.process.ProcessException;
 import com.walmartlabs.concord.server.process.ProcessKey;