+++ server/impl/src/main/java/com/walmartlabs/concord/server/security/github/GithubRealm.java
@@ -20,7 +20,7 @@ package com.walmartlabs.concord.server.security.github;
  * =====
  */
 
-import com.walmartlabs.concord.server.metrics.WithTimer;
+import com.walmartlabs.concord.server.sdk.metrics.WithTimer;
 import com.walmartlabs.concord.server.security.PrincipalUtils;
 import com.walmartlabs.concord.server.security.UserPrincipal;
 import com.walmartlabs.concord.server.user.UserManager;