+++ server/impl/src/main/java/com/walmartlabs/concord/server/websocket/WebSocketMetricsModule.java
@@ -23,7 +23,7 @@ package com.walmartlabs.concord.server.websocket;
 import com.codahale.metrics.Gauge;
 import com.google.inject.AbstractModule;
 import com.google.inject.multibindings.Multibinder;
-import com.walmartlabs.concord.server.metrics.GaugeProvider;
+import com.walmartlabs.concord.server.sdk.metrics.GaugeProvider;
 
 import javax.inject.Named;
 import javax.inject.Provider;