+++ server/sdk/src/main/java/com/walmartlabs/concord/server/sdk/metrics/GaugeProvider.java
@@ -1,4 +1,4 @@
-package com.walmartlabs.concord.server.metrics;
+package com.walmartlabs.concord.server.sdk.metrics;
 
 /*-
  * *****